Job Request - legal confirmtion of solicitor's job

+++ FintrakBanking.ViewModels/WorkFlow/JobRequestViewModel.cs
@@ -123,7 +123,7 @@ namespace FintrakBanking.ViewModels.WorkFlow
         public short collateralStateId { get; set; }
         public string requestCode  { get; set; }
         public int operationId { get; set; }
-        public int? currencyId { get; set; }
+        public short? currencyId { get; set; }
         public bool isInitiation { get; set; }
     }
